Add clock delay parameter for emulated devices (e.g. HH-12INC)
This patch adds a clock delay parameter (_clockDelay) to the constructor,
which indicates the desired delay in microseconds between clock cycles
when reading the sensor device.

This is required for the HH-12INC inclinometer by DF1SR and SQ6EDU.

According to VK5DJ, this inclinometer emulates an AS5045 via an ARM STM32F,
and needs a slower clock rate (https://www.vk5dj.com/page63.html).

I've found that a value of 30us works for my setup.

Example use:

  static AS5045 my_as5405(CS_pin, CLK_pin, DO_pin, 0xFF, 30);

The default value is zero.

Signed-off-by: James Morris W7TXT <jmorris@namei.org>